Fix default values and move normalizeSelection out of the component
