+++ src/gm3/query/util.js
@@ -27,4 +27,28 @@ export const applyPixelTolerance = (queryFeature, querySource, resolution, defau
     return queryFeature;
 };
 
+export const normalizeSelection = (selectionFeatures) => {
+    // OpenLayers handles MultiPoint geometries in an awkward way,
+    // each feature is a 'MultiPoint' type but only contains one feature,
+    //  this normalizes that in order to be submitted properly to query services.
+    if (selectionFeatures && selectionFeatures.length > 0) {
+        if (selectionFeatures[0].geometry.type === 'MultiPoint') {
+            const all_coords = [];
+            selectionFeatures.forEach(feature => {
+                if(feature.geometry.type === 'MultiPoint') {
+                    all_coords.push(feature.geometry.coordinates[0]);
+                }
+            });
+            return [{
+                type: 'Feature',
+                properties: {},
+                geometry: {
+                    type: 'MultiPoint',
+                    coordinates: all_coords
+                }
+            }];
+        }
+    }
+    return selectionFeatures;
+};
